requirements.txt
livekit-agents>=0.8.0
livekit-api>=0.11.0
livekit-rtc>=0.10.0
deepgram-sdk>=3.0.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
httpx>=0.25.0
aiofiles>=23.0.0
twilio>=8.0.0
websockets>=12.0
aiohttp>=3.9.0